// frontend/book-review-frontend/src/app/models/auth.model.ts
export interface User {
  id: string;
  userName: string;
  email: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  confirmPassword: string;
}

export interface AuthResponse {
  token: string;
  expiration: string;
  userId: string;
  userName: string;
  email: string;
}